rename file, fixed styles
