+++ src/modules/dashboard/components/Incorporation/DescriptionForm/DescriptionForm.css
@@ -1,5 +1,4 @@
 .container {
-  padding-right: 37px;
   padding-bottom: 50px;
   width: 100%;
 }
