+++ src/modules/pages/components/IncorporationPage/IncorporationPage.css.d.ts
@@ -3,7 +3,5 @@ export const navBarHeight: string;
 export const main: string;
 export const mainContainer: string;
 export const mainContent: string;
-export const titleCommentsContainer: string;
 export const sidebar: string;
-export const wrapper: string;
 export const spinnerContainer: string;
