+++ src/modules/pages/components/IncorporationPage/IncorporationPage.tsx
@@ -5,8 +5,8 @@ import { Formik } from 'formik';
 import { useColonyFromNameQuery } from '~data/generated';
 import { getMainClasses } from '~utils/css';
 import { SpinnerLoader } from '~core/Preloaders';
-import IncorporationForm from '~dashboard/DAOIncorporation/IncorporationForm';
-import DescriptionForm from '~dashboard/DAOIncorporation/DescriptionForm';
+import IncorporationForm from '~dashboard/Incorporation/IncorporationForm';
+import DescriptionForm from '~dashboard/Incorporation/DescriptionForm';
 
 import { initialValues } from './constants';
 import styles from './IncorporationPage.css';
